# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DirectorOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssetAssetType } from "./assetAssetType";
import type { AssetMetadata } from "./assetMetadata";

export interface Asset {
  id: number;
  projectId: number;
  assetType: AssetAssetType;
  storageUri: string;
  thumbnailUri?: string | null;
  width?: number | null;
  height?: number | null;
  durationMs?: number | null;
  metadata?: AssetMetadata;
  createdAt: Date;
}