# lib/api-zod/src/generated/types/createProjectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DirectorOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectInput {
  name: string;
  description?: string | null;
  aspectRatio?: string | null;
  targetDuration?: number | null;
}