# lib/api-zod/src/generated/types/createSceneInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DirectorOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSceneInput {
  name: string;
  orderIndex: number;
  summary?: string | null;
  startTimeMs?: number | null;
  endTimeMs?: number | null;
}